GLcell and GLview: minor modifications for Visual Studio

+++ moose/branches/3D-viz/gl/src/GLcell.h
@@ -90,7 +90,7 @@ class GLcell
 	double highValue_;
 	double lowValue_;
 
-	unsigned int testTicker_; // used by testInsertVecParticleData()
+	double testTicker_; // used by testInsertVecParticleData()
 
 	vector< Id > renderList_;
 	vector< ParticleData > vecParticleData_;
